Add work-in-progress CUDL item schema
This is a JSON schema for CUDL Viewer items. The flexible nature of
the CUDL item JSON means that a schema can't really constrain it much,
as the main metadata is essentially arbitary JSON where meaning is
given to properties that follow a pattern, and anything else is
ignored. So it serves more as documentation than a means of detecting
errors.