Tweak example data to be closer to JSON-LD
URI references that we'll resolve at build time are now consistently
using { "@id": "uri/value" }. Previously I was relying on context to
decide what was a resolvable URI and what was just a string property.
Using this @id syntax is slightly more verbose, but has a few
advantages:

1. It allows URIs to be consistently represented pre and post loading/
  transforming. Previously I was thinking that we'd have a
  transformation rule like:
    "thing": { "href": "./url" } -> "thing": "/abs/url"

  This results in some information loss, as we need out of band context
  to know that the value of "thing" is a URI reference. This makes it
  harder to generically handle our JSON data.
2. (as implied by 1.) It allows us to have one rule for determining
  which string values to resolve to treat as URIs and resolve to other
  files at build time. This is simpler to implement, and simpler for
  humans.
2. It's consistent with JSON-LD.
  My goal is to move to building the whole dataset as linked data via
  JSON-LD, using it's intrinsic support for URIs to guide what's a
  reference to other files. We are not yet at that point, and our JSON
  files are just normal JSON. However this gets us part of the way
  there, which I hope will make transitioning easier.

When you see a property like "thing": { "@id": "./foo/bar" } you can
read the right hand side object value as being a URI reference
(./foo/bar). In the same way as when you see "thing": 42 you know the
value is a number.

With JSON-LD you can specify additional properties for a node
identified in this way, e.g:
    "thing": { "@id": "./foo/bar", "name": "frob" }
At the moment we shouldn't do this, because we don't have a good way to
associate the additional "name" property with ./foo/bar, because we're
not actually using a linked-data model. So we should restrict ourselves
to referencing nodes, not declaring new data about them.

Instead we can do things like:
    {
        "thing": { "@id": "./foo/bar" },
        "name": "frob"
    }
I.e. wrap an extra layer around ./foo/bar rather than attaching data
to it.