Restructure items to support flexible associated data
Items now have a general-purpose "data" section, which supports
associating typed data to items, along with role tags to specify how
the data relates to the item and or how it should be used.

Currently two types of data objects are defined:
  - A properties list which replaces the top-level item "properties"
    section.
  - A link, associating an external entity to the item via a URL. The
    @role(s) of the link are used to specify what the link is and how
    it should be used. e.g. TEI metadata for an item might be linked
    using with the cdl-role:descriptive-metadata

Another type of data will be defined imminently - data defining fields
values to use when building the search index.